Keep IDF lib copy logic in espidf.py
